requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
autogluon.tabular[fastai]==1.5.0
ipython
catboost
lightgbm
pandas>=2.2.3
numpy<2
pyyaml
cbbd
requests
